Refactor: Move `_get_ts2mp4_version` to `__init__.py` and its tests to `test_init.py`
Moved the `_get_ts2mp4_version` utility function from `ts2mp4.py` to `ts2mp4/__init__.py` to centralize package-level metadata retrieval.

Correspondingly, its unit tests were moved from `tests/test_ts2mp4.py` to a new dedicated test file `tests/test_init.py`.

This improves modularity and ensures that tests for package-level utilities are located appropriately.
